requirements.txt
pandas==1.5.3
scikit-learn==1.2.2
transformers==4.30.2
torch==2.0.1
fastapi==0.101.0
uvicorn==0.23.2
nltk==3.8.1
rouge-score==0.1.2
language-tool-python==2.7.1
python-dotenv==1.0.0